fix(ui): break chat sidebar import cycle
